refactor(anyharness): split terminal domain and live runtime
