+++ anyharness/crates/anyharness-lib/src/app/mod.rs
@@ -30,8 +30,9 @@ use crate::domains::reviews::runtime::ReviewRuntime;
 use crate::domains::reviews::service::ReviewService;
 use crate::domains::reviews::store::{ReviewDeleteParticipant, ReviewStore};
 use crate::domains::runtime_config::service::{RuntimeConfigService, RuntimeConfigStore};
-use crate::domains::runtime_config::session_extension::RuntimeConfigSessionLaunchExtension;
+use crate::domains::terminals::store::TerminalStore;
 use crate::live::sessions::LiveSessionManager;
+use crate::live::terminals::TerminalService;
 use crate::persistence::Db;
 use crate::repo_roots::service::RepoRootService;
 use crate::repo_roots::store::RepoRootStore;
@@ -51,8 +52,6 @@ use crate::sessions::subagents::mcp::auth::SubagentMcpAuth;
 use crate::sessions::subagents::service::SubagentService;
 use crate::sessions::subagents::store::SubagentStore;
 use crate::sessions::workspace_naming::mcp::auth::WorkspaceNamingMcpAuth;
-use crate::terminals::store::TerminalStore;
-use crate::terminals::TerminalService;
 use crate::workspaces::access_gate::WorkspaceAccessGate;
 use crate::workspaces::access_store::WorkspaceAccessStore;
 use crate::workspaces::checkout_gate::CheckoutDeletionGate;
