+++ anyharness/crates/anyharness-lib/src/domains/mod.rs
@@ -5,3 +5,4 @@ pub mod plans;
 pub mod plugins;
 pub mod reviews;
 pub mod runtime_config;
+pub mod terminals;
