+++ anyharness/crates/anyharness-lib/src/domains/terminals/model.rs
