+++ anyharness/crates/anyharness-lib/src/domains/terminals/store.rs
