+++ anyharness/crates/anyharness-lib/src/live/mod.rs
@@ -1,1 +1,2 @@
 pub mod sessions;
+pub mod terminals;
