+++ anyharness/crates/anyharness-lib/src/live/terminals/mod.rs
@@ -1,0 +1,11 @@
+mod driver;
+mod handle;
+mod manager;
+mod output_sink;
+mod pty_command;
+mod replay;
+mod setup_process;
+mod shell;
+
+pub use manager::TerminalService;
+pub use output_sink::TerminalOutputEvent;
