--- anyharness/crates/anyharness-lib/src/terminals/mod.rs
@@ -1,7 +1,0 @@
-mod shell;
-pub mod store;
-
-pub mod model;
-pub mod service;
-
-pub use service::TerminalService;
